requirements.txt
python-docx==0.8.11
beautifulsoup4==4.12.2
requests==2.31.0
difflib==0.0.1 